requirements.txt
# Core
streamlit==1.28.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Data
pandas==2.1.4
numpy==1.26.3
polars==0.19.19

# Databricks
databricks-sql-connector==3.0.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
alembic==1.13.1

# Azure
azure-identity==1.14.0
azure-storage-blob==12.19.0
azure-keyvault-secrets==4.7.0
azure-ai-openai==1.3.0
azure-cognitiveservices-speech==1.35.0
openai>=1.0.0

# ML & Recommendations
scikit-learn==1.3.2
xgboost==2.0.2

# Data Generation (Synthetic)
faker==20.1.0
sdv==1.3.0
pydbgen==0.2.11

# Utils
python-dateutil==2.8.2
requests==2.31.0
Pillow==10.1.0
PyPDF2==3.17.1
python-docx==0.8.11
pdfplumber==0.10.3
qrcode==7.4.2

# Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Logging
python-json-logger==2.0.7
